chore: breakup account form
